readformatlolo editcm is antiquated
basic_data_tables creates some basic data tables.

Adding readformat_pgp_recentlymeasured. - this reads-in all of the data for stands that have recent measurement data. Does not include actual recent data - still waiting on that.

modified species_comp_functions.

Created folder which contains lolo data with missing distances/azimuth. This was in the lolo data file and making read-in activities difficult. not sure if this is needed in lolo dataset. If so, we can add it into the readformat_pgp_recentlymeasured file.